refactor(langy): the package sorts itself — event-sourcing under one roof
Pure restructure ahead of the blocks kernel (ADR-060): the shared package's
nine flat modules become a tree that says what each thing is. json.ts (the
shared JSON value schema, used by event contracts and soon the blocks) stays
at the root beside constants.ts; everything conversation-event-sourced moves
under event-sourcing/ — contracts/ for the cross-runtime wire shapes (cursor,
events, and the turn wire envelope newly extracted from the fold; the fold
keeps langyTurnToolCallSchema, which must stay composed inside the package's
own zod instance), folds/ for the pure reducers, and the browser-side
machines (turnPhase, turnProjection) directly at the event-sourcing root.
Tests move beside their subjects, the barrel remains the only public API, and
sideEffects:false lets consumers tree-shake a package that is now honestly
side-effect-free. Zero files outside the package touched.
