+++ packages/langy/src/event-sourcing/contracts/cursor.ts
