+++ packages/langy/src/index.ts
@@ -1,9 +1,10 @@
 export * from "./constants";
-export * from "./shared";
-export * from "./events";
-export * from "./cursor";
-export * from "./turnPhase";
-export * from "./turnFold";
-export * from "./turnProjection";
-export * from "./conversationFold";
-export * from "./messageMap";
+export * from "./json";
+export * from "./event-sourcing/contracts/events";
+export * from "./event-sourcing/contracts/cursor";
+export * from "./event-sourcing/contracts/turnWire";
+export * from "./event-sourcing/turnPhase";
+export * from "./event-sourcing/turnProjection";
+export * from "./event-sourcing/folds/turnFold";
+export * from "./event-sourcing/folds/conversationFold";
+export * from "./event-sourcing/folds/messageMap";
